moved functions into utils
